requirements.txt
Django==2.2.5
django-crispy-forms==1.7.2
django-dynamic-preferences==1.7.1
persisting-theory==0.2.1
Pillow==6.2.0
python-slugify==3.0.4
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
